refactor: move model-listing files to a separate directory
