+++ packages/runtime/src/index.ts
@@ -1,29 +1,8 @@
 // Copyright (c) 2020-2022 Climate Interactive / New Venture Fund
 
 export * from './_shared'
-export * from './runnable-model'
+export * from './model-listing'
 export * from './js-model'
 export * from './wasm-model'
 export * from './model-runner'
 export * from './model-scheduler'
-
-import type { RunnableModel } from './runnable-model'
-import { type JsModel, initJsModel } from './js-model'
-import { type WasmModule, initWasmModel } from './wasm-model'
-
-/**
- * Create a `RunnableModel` from a given `JsModel` or `WasmModule` that was generated by the
- * SDEverywhere transpiler/builder.
- */
-export function createRunnableModel(generatedModel: JsModel | WasmModule): RunnableModel {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  if ((generatedModel as any)['setModelFunctions'] !== undefined) {
-    // Assume it's a `JsModel`
-    return initJsModel(generatedModel as JsModel)
-  } else {
-    // Otherwise, assume it's a `WasmModule`
-    // TODO: Remove the second parameter once plugin-wasm is updated to include
-    // outputVarIds in the generated WasmModule
-    return initWasmModel(generatedModel as WasmModule, [])
-  }
-}
