+++ packages/runtime/src/model-runner/index.ts
@@ -2,6 +2,3 @@
 
 export * from './model-runner'
 export * from './synchronous-model-runner'
-
-/** @hidden This is not part of the public API; exposed only for use in testing tools. */
-export * from './model-listing'
